extract collection-specific settings into separate pydantic model
